% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-db.R
\name{alter_table_schemas}
\alias{alter_table_schemas}
\title{Title}
\usage{
alter_table_schemas(.con, from_table_names, new_schema)
}
\arguments{
\item{.con}{connection}

\item{new_schema}{new schema name}

\item{old_schemapreviou}{shema name}

\item{table_name}{table name}
}
\description{
Title
}
