% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{check_resp_status_and_tidy_response}
\alias{check_resp_status_and_tidy_response}
\title{Check resp status and tidy output to a tibble}
\usage{
check_resp_status_and_tidy_response(
  resp,
  json_response,
  column_name1,
  column_name2
)
}
\arguments{
\item{resp}{response code}

\item{json_response}{response object}

\item{column_name1}{first column of name of response object}

\item{column_name2}{second column of name of response object}
}
\value{
tibble
}
\description{
Check resp status and tidy output to a tibble
}
