% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{configure_md_user_settings}
\alias{configure_md_user_settings}
\title{Configure motherduck user's settings}
\usage{
configure_md_user_settings(
  user_name,
  motherduck_token = "MOTHERDUCK_TOKEN",
  token_type = "read_write",
  instance_size = "pulse",
  flock_size = 0
)
}
\arguments{
\item{user_name}{motherduck user name}

\item{motherduck_token}{user's access token}

\item{token_type}{the to be token type}

\item{instance_size}{the to be instance size}

\item{flock_size}{the to be flock size}
}
\value{
tibble
}
\description{
Configure motherduck user's settings
}
