% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-md.R
\name{connect_to_motherduck}
\alias{connect_to_motherduck}
\title{Create connection to motherduck}
\usage{
connect_to_motherduck(motherduck_token)
}
\arguments{
\item{motherduck_token}{motherduck token saved in your environment file}
}
\value{
connection
}
\description{
creates a DuckDB connection, installs and loads the motherduck extension and finally
executes  `DBI::dbExecute(con, "PRAGMA MD_CONNECT")` to connect to motherduck through your
motherduck token
}
