% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-db.R
\name{convert_table_to_sql_id}
\alias{convert_table_to_sql_id}
\title{Convert table to SQL IDs}
\usage{
convert_table_to_sql_id(x)
}
\arguments{
\item{x}{tibble of database, schemas and tables}
}
\value{
list of sql IDs
}
\description{
Convert table to SQL IDs
}
