% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{convert_to_seconds}
\alias{convert_to_seconds}
\title{Convert units to seconds}
\usage{
convert_to_seconds(number, units)
}
\arguments{
\item{number}{numeriv value}

\item{units}{second, minute, day, month, year or never}
}
\value{
number
}
\description{
Convert units to seconds
}
\examples{
convert_to_seconds(300,"days")
}
