% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-db.R
\name{copy_tables_to_new_location}
\alias{copy_tables_to_new_location}
\title{Copy tables to new schema}
\usage{
copy_tables_to_new_location(
  .con,
  from_table_names,
  to_database_name,
  to_schema_name
)
}
\arguments{
\item{.con}{connection}

\item{from_table_names}{tibble of tables to be copied with `database_name`,`schema_name` and `table_name` columns}

\item{to_database_name}{target database name}

\item{to_schema_name}{target schema_name}
}
\value{
message
}
\description{
Copy tables to new schema
}
