% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/minimumDistanceMethod.R
\name{counter}
\alias{counter}
\title{Get 0-1 vector which is used for calculating empirical measure}
\usage{
counter(e2, box2)
}
\arguments{
\item{e2}{integer}

\item{box2}{matrix which has 0-1 elements related to Scheffe set}
}
\value{
0-1 vector
}
\description{
Get 0-1 vector which is used for calculating empirical measure
}

