% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{create_if_not_exist_motherduck_share}
\alias{create_if_not_exist_motherduck_share}
\title{Title}
\usage{
create_if_not_exist_motherduck_share(
  .con,
  share_name,
  database_name,
  access = "PUBLIC",
  visibility = "LISTED",
  update = "AUTOMATIC"
)
}
\arguments{
\item{update}{}
}
\description{
Title
}
