% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{create_if_not_exists_share}
\alias{create_if_not_exists_share}
\title{Create a MD share of a database}
\usage{
create_if_not_exists_share(
  .con,
  share_name,
  database_name,
  access = "PUBLIC",
  visibility = "LISTED",
  update = "AUTOMATIC"
)
}
\arguments{
\item{.con}{MD connection}

\item{share_name}{new share name}

\item{database_name}{target database}

\item{access}{"RESTRICTED" or "PUBLIC"}

\item{visibility}{"HIDDEN" or "LISTED"}

\item{update}{"AUTOMATIC" or "MANUAL"}
}
\value{
message
}
\description{
Create a MD share of a database
}
