% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{create_md_access_token}
\alias{create_md_access_token}
\title{Create a motherduck access token}
\usage{
create_md_access_token(
  user_name,
  token_type,
  token_name,
  token_expiration_number,
  token_expiration_unit,
  motherduck_token = "MOTHERDUCK_TOKEN"
)
}
\arguments{
\item{user_name}{new users name}

\item{token_type}{the token type}

\item{token_name}{the token name}

\item{token_expiration_number}{the token expiration number - minimum 300 seconds}

\item{token_expiration_unit}{the token expiration unit (\code{seconds},\code{minutes},\code{days},\code{weeks},\code{months},\code{years},\code{never})}

\item{motherduck_token}{admin user's token}
}
\value{
tibble
}
\description{
Create a motherduck access token
}
