% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{create_md_user}
\alias{create_md_user}
\title{Create a new motherduck token}
\usage{
create_md_user(user_name, motherduck_token = "MOTHERDUCK_TOKEN")
}
\arguments{
\item{user_name}{new motherduck user name}

\item{motherduck_token}{admin user's token}
}
\value{
tibble
}
\description{
Create a new motherduck token
}
