% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-db.R
\name{create_or_replace_database}
\alias{create_or_replace_database}
\title{Create or replace database}
\usage{
create_or_replace_database(.con, database_name)
}
\arguments{
\item{.con}{connection}

\item{database_name}{new database name}
}
\value{
message
}
\description{
Create or replace database
}
