% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{create_or_replace_share}
\alias{create_or_replace_share}
\title{@title Create or replace a motherduck share}
\usage{
create_or_replace_share(
  .con,
  share_name,
  database_name,
  access = "PUBLIC",
  visibility = "LISTED",
  update = "AUTOMATIC"
)
}
\arguments{
\item{.con}{md connection}

\item{share_name}{shame name}

\item{database_name}{database name to be shared}

\item{access}{either "RESTRICTED" or "PUBLIC"}

\item{visibility}{either "HIDDEN"or  "LISTED"}

\item{update}{either "AUTOMATIC" or "MANUAL"}
}
\value{
message
}
\description{
@title Create or replace a motherduck share
}
