% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-db.R
\name{create_table}
\alias{create_table}
\alias{create_table_tbl}
\title{Overwrite or append tibble to database}
\usage{
create_table_tbl(
  .data,
  .con,
  database_name,
  schema_name,
  table_name,
  write_type = "overwrite"
)

create_table(
  .data,
  .con,
  database_name,
  schema_name,
  table_name,
  write_type = "overwrite"
)
}
\arguments{
\item{.data}{tibble or dbi object}

\item{.con}{duckdb or MD connection}

\item{database_name}{database name}

\item{schema_name}{schema name}

\item{table_name}{table name}

\item{write_type}{overwrite or append}
}
\value{
nothing

message
}
\description{
Overwrite or append tibble to database

Title
}
