% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-db.R
\name{create_table_dbi}
\alias{create_table_dbi}
\title{Create table from DBI object}
\usage{
create_table_dbi(
  .data,
  .con,
  database_name,
  schema_name,
  table_name,
  write_type = "overwrite"
)
}
\arguments{
\item{.data}{dbi object}

\item{.con}{connection}

\item{database_name}{database name}

\item{schema_name}{schema name}

\item{table_name}{table name}

\item{write_type}{overwrite or append}
}
\value{
message
}
\description{
Create table from DBI object
}
