% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-db.R
\name{delete_database}
\alias{delete_database}
\title{Title}
\usage{
delete_database(.con, database_name)
}
\arguments{
\item{.con}{connection}

\item{database_name}{database name}
}
\value{
message
}
\description{
Title
}
