% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{delete_md_access_token}
\alias{delete_md_access_token}
\title{Title}
\usage{
delete_md_access_token(
  user_name,
  token_name,
  motherduck_token = "MOTHERDUCK_TOKEN"
)
}
\arguments{
\item{user_name}{motherduck user name}

\item{token_name}{motherduck token name}

\item{motherduck_token}{admin user's token}
}
\value{
tibble
}
\description{
Title
}
