% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-db.R
\name{delete_schema}
\alias{delete_schema}
\title{Title}
\usage{
delete_schema(.con, database_name, schema_name, cascade = TRUE)
}
\arguments{
\item{.con}{connection}

\item{database_name}{database name}

\item{schema_name}{schema name}

\item{cascade}{}
}
\value{
message
}
\description{
Title
}
