% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/minimumDistanceMethod.R
\name{deltaboxm}
\alias{deltaboxm}
\title{Calculating delta}
\usage{
deltaboxm(ij, box, box2, mu_box2, grid)
}
\arguments{
\item{ij}{number order vector}

\item{box}{matrix which has estimated values of all
kernel density estimators}

\item{box2}{matrix which has 0-1 elements related to Scheffe set}

\item{mu_box2}{matrix which has values of all empirical measures}

\item{grid}{length of grid in domain}
}
\value{
delta value
}
\description{
Calculating delta
}

