% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{describe_share}
\alias{describe_share}
\title{Describe share}
\usage{
describe_share(.con, share_path)
}
\arguments{
\item{.con}{connection}

\item{share_path}{shared path name}
}
\value{
tibble
}
\description{
Describe share
}
