% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/minimumDistanceMethod.R
\name{dmm}
\alias{dmm}
\title{Get number order matrix which is used in md}
\usage{
dmm(i, prod2)
}
\arguments{
\item{i}{integer}

\item{prod2}{integer which is defined in md}
}
\value{
matrix related to number order
}
\description{
Get number order matrix which is used in md
}

