% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/minimumDistanceMethod.R
\name{fhat}
\alias{fhat}
\title{Calculating estimated density value on some x with bandwidth h}
\usage{
fhat(x, h, data)
}
\arguments{
\item{x}{scalar}

\item{h}{bandwidth}

\item{data}{data sample}
}
\value{
estimated density value
}
\description{
Calculating estimated density value on some x with bandwidth h
}
\examples{
fhat(0,0.2,rnorm(100))
}

