% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/minimumDistanceMethod.R
\name{fhatboxm}
\alias{fhatboxm}
\title{Get estimated values of kernel density estimator on domain}
\usage{
fhatboxm(ij, data_for_d, h, x)
}
\arguments{
\item{ij}{number order vector}

\item{data_for_d}{data sample which is split to be used for
kernel density estimator}

\item{h}{bandwidth}

\item{x}{scalar}
}
\value{
estimated values of kernel density estimator on domain
}
\description{
Get estimated values of kernel density estimator on domain
}

