% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-md.R
\name{list_database}
\alias{list_database}
\title{list database objects}
\usage{
list_database(.con)
}
\arguments{
\item{.con}{connection}

\item{type}{'database', 'schema' or 'views'}
}
\value{
tibble
}
\description{
list database objects
}
\examples{
#' con <- DBI::dbConnect(duckdb::duckdb())
lmd(con,type='database')
}
