% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-md.R
\name{list_fns}
\alias{list_fns}
\title{List database functions}
\usage{
list_fns(.con)
}
\arguments{
\item{.con}{connection}
}
\value{
tibble
}
\description{
List database functions
}
\examples{
con <- DBI::dbConnect(duckdb::duckdb())
list_db_fns(con)
}
