% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts_token.R
\name{list_md_active_accounts}
\alias{list_md_active_accounts}
\title{Title}
\usage{
list_md_active_accounts(motherduck_token = "MOTHERDUCK_TOKEN")
}
\arguments{
\item{motherduck_token}{}
}
\value{
list
}
\description{
Title
}
