% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts_token.R
\name{list_md_user_instance}
\alias{list_md_user_instance}
\title{Title}
\usage{
list_md_user_instance(user_name, motherduck_token = "MOTHERDUCK_TOKEN")
}
\arguments{
\item{user_name}{mother duck user name}

\item{motherduck_token}{mother duck token or environment name}
}
\description{
Title
}
