% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts_token.R
\name{list_md_user_tokens}
\alias{list_md_user_tokens}
\title{Title}
\usage{
list_md_user_tokens(user_name, motherduck_token = "MOTHERDUCK_TOKEN")
}
\arguments{
\item{user_name}{motherduck user name}

\item{motherduck_token}{motherduck token or environment nick name}

\item{.con}{duckdb connection}
}
\value{
tibble
}
\description{
Title
}
