% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{list_owned_shares}
\alias{list_owned_shares}
\title{List all owned shares}
\usage{
list_owned_shares(.con)
}
\arguments{
\item{.con}{motherduck connection}
}
\value{
tibble
}
\description{
List all owned shares
}
