% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-md.R
\name{list_settings}
\alias{list_settings}
\alias{list_setting}
\title{List database settings}
\usage{
list_setting(.con)
}
\arguments{
\item{.con}{dubdb or md connection}
}
\value{
tibble
}
\description{
List database settings
}
