% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-md.R
\name{load_extensions}
\alias{load_extensions}
\title{Loand (and install) motherduck extensions}
\usage{
load_extensions(.con, extension_names)
}
\arguments{
\item{.con}{duckdb connection}

\item{extension_names}{DuckDB extension names}

\item{silent_msg}{indicate if you want a success / failure report after installation and loading}
}
\value{
message
}
\description{
Installs and loads valid DuckDB extensions
}
\examples{
con <- DBI::dbConnect(duckdb::duckdb())
load_extensions(con,'motherduck')

}
