% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-md.R
\name{ls.Pool}
\alias{ls.Pool}
\title{list database objects}
\usage{
\method{ls}{Pool}(.con)
}
\arguments{
\item{.con}{connection}
}
\value{
tibble
}
\description{
list database objects
}
\examples{
#' con <- DBI::dbConnect(duckdb::duckdb())
list_db_fns(con)
}
