% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-md.R
\name{pwd}
\alias{pwd}
\title{Print current databases}
\usage{
pwd(.con)
}
\arguments{
\item{.con}{}
}
\value{
tibble
}
\description{
Prints the current database that you are in (adopts language from linux)
}
\examples{
#' con <- DBI::dbConnect(duckdb::duckdb())
pwd(con)
}
