% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-md.R
\name{read_httpfs}
\alias{read_httpfs}
\title{read httpfs files}
\usage{
read_httpfs(.con, file_path)
}
\arguments{
\item{.con}{DuckDB connection}

\item{file_path}{to httpfs files}
}
\value{
message
}
\description{
Enables reading of httpfs files
}
