% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-md.R
\name{read_parquet}
\alias{read_parquet}
\title{read httpfs files}
\usage{
read_parquet(.con, file_path)
}
\arguments{
\item{.con}{connection}

\item{file_path}{file path to parquet files}
}
\value{
message
}
\description{
read httpfs files
}
