% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fns.R
\name{read_parquet_duckdb}
\alias{read_parquet_duckdb}
\title{Title}
\usage{
read_parquet_duckdb(
  .con,
  to_database_name,
  to_schema_name,
  to_table_name,
  file_path
)
}
\arguments{
\item{empty_as_varchar}{}
}
\description{
Title
}
