% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/minimumDistanceMethod.R
\name{scheffe}
\alias{scheffe}
\title{Calculating Scheffe sets}
\usage{
scheffe(ij, box)
}
\arguments{
\item{ij}{number order vector}

\item{box}{estimated values of all kernel density estimators}
}
\value{
0-1 vector
}
\description{
Calculating Scheffe sets
}

