% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{show_current_user}
\alias{show_current_user}
\title{Title}
\usage{
show_current_user(motherduck_token = "MOTHERDUCK_TOKEN", return = "msg")
}
\arguments{
\item{.con}{duckdb connectoin}
}
\value{
tibble
}
\description{
Title
}
