% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-md.R
\name{show_duckdb_settings}
\alias{show_duckdb_settings}
\title{Show DuckDB settings}
\usage{
show_duckdb_settings(.con)
}
\arguments{
\item{.con}{connection}
}
\value{
tibble
}
\description{
Show DuckDB settings
}
\examples{
con <- DBI::dbConnect(duckdb::duckdb())
show_duckdb_settings(con)
}
