% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-md.R
\name{show_motherduck_token}
\alias{show_motherduck_token}
\title{Show your motherduck token}
\usage{
show_motherduck_token(.con)
}
\arguments{
\item{.con}{connection}
}
\value{
message
}
\description{
Show your motherduck token
}
