% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/minimumDistanceMethod.R
\name{subcounter}
\alias{subcounter}
\title{Auxiliary function which is used in md}
\usage{
subcounter(l, data, x)
}
\arguments{
\item{l}{integer}

\item{data}{data sample}

\item{x}{scalar}
}
\value{
integer
}
\description{
Auxiliary function which is used in md
}

