% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-md.R
\name{summarise.tbl_lazy}
\alias{summarise.tbl_lazy}
\title{Summarize for DBI objects}
\usage{
\method{summarise}{tbl_lazy}(.data)
}
\arguments{
\item{.data}{dbi object}
}
\value{
DBI object
}
\description{
Summarize for DBI objects
}
