% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-md.R
\name{summarize.tbl_lazy}
\alias{summarize.tbl_lazy}
\title{Summarize for DBI objects}
\usage{
\method{summarize}{tbl_lazy}(.data)
}
\arguments{
\item{.data}{dbi object}
}
\value{
DBI object
}
\description{
Summarize for DBI objects
}
