% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-md.R
\name{summary.pool}
\alias{summary.pool}
\title{Summarize for DBI objects}
\usage{
\method{summary}{pool}(object, ...)
}
\arguments{
\item{.data}{dbi object}
}
\value{
DBI object
}
\description{
Summarize for DBI objects
}
