% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-db.R
\name{upload_database_to_md}
\alias{upload_database_to_md}
\title{Upload a local database to motherduck}
\usage{
upload_database_to_md(.con, from_db_name, to_db_name)
}
\arguments{
\item{.con}{motherduck connection}

\item{from_db_name}{the local database to be copied}

\item{to_db_name}{the name of the motherduck database to be created}
}
\value{
print statement
}
\description{
Upload a local database to motherduck
}
