% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_fns.R
\name{validate_and_print_database_loction}
\alias{validate_and_print_database_loction}
\title{Title}
\usage{
validate_and_print_database_loction(.con)
}
\arguments{
\item{.con}{connection}
}
\value{
print message
}
\description{
Title
}
