% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts_token.R
\name{validate_and_show_current_user}
\alias{validate_and_show_current_user}
\title{Title}
\usage{
validate_and_show_current_user(motherduck_token, user_name, return = "msg")
}
\arguments{
\item{.con}{duckdb connectoin}
}
\value{
tibble
}
\description{
Title
}
