% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-md.R
\name{validate_connection_status}
\alias{validate_connection_status}
\title{Validate Mother Duck Connection Status}
\usage{
validate_connection_status(.con, return_type = "msg")
}
\arguments{
\item{.con}{connection}

\item{return_type}{return message or logical value of connection status}
}
\value{
confirmation or warning message
}
\description{
Validate Mother Duck Connection Status
}
\examples{
con <- DBI::dbConnect(duckdb::duckdb())
validate_connection_status(con)
}
