% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-md.R
\name{validate_extension_install_status}
\alias{validate_extension_install_status}
\title{Validate that the Motherduck extension correctly loaded}
\usage{
validate_extension_install_status(.con, extension_names, return_type = "msg")
}
\arguments{
\item{.con}{connection obj}

\item{extension_names}{list of extension names that you want to validate}

\item{return_type}{'msg' or 'ext'}
}
\value{
message or extension names
}
\description{
Validate that the Motherduck extension correctly loaded
}
\examples{
\dontrun{
library(DBI)
con <- dbConnect(duckdb::duckdb())
validate_extension_install_status(con,extension_names=c('excelA','arrow'),return_type='ext')
}
}
