% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-md.R
\name{validate_extension_load_status}
\alias{validate_extension_load_status}
\title{Validate  Motherduck extensions are correctly loaded}
\usage{
validate_extension_load_status(.con, extension_names, return_type = "msg")
}
\arguments{
\item{.con}{connection obj}

\item{extension_names}{list of extension names that you want to validate}

\item{return_type}{'msg' or 'ext'}
}
\value{
message or extension names
}
\description{
Validate  Motherduck extensions are correctly loaded
}
\examples{
\dontrun{
con <- DBI::dbConnect(duckdb::duckdb())
validate_extension_load_status(con,extension_names=c('excel','arrow'),return_type='ext')
}
}
