% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{validate_flock_size}
\alias{validate_flock_size}
\title{Validate flock size args}
\usage{
validate_flock_size(flock_size)
}
\arguments{
\item{flock_size}{whole number between 0-60}
}
\value{
integer
}
\description{
Validate flock size args
}
