% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{validate_instance_size}
\alias{validate_instance_size}
\title{Validate instance size args}
\usage{
validate_instance_size(instance_size)
}
\arguments{
\item{instance_size}{select either "pulse", "standard", "jumbo", "mega", "giga"}
}
\value{
character vector
}
\description{
Validate instance size args
}
