% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{validate_motherduck_token_env}
\alias{validate_motherduck_token_env}
\title{Validate if motherduck token in Environment file}
\usage{
validate_motherduck_token_env(motherduck_token = "MOTHERDUCK_TOKEN")
}
\arguments{
\item{motherduck_token}{}
}
\value{
character vector
}
\description{
Validate if motherduck token in Environment file
}
