% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{validate_token_type}
\alias{validate_token_type}
\title{Validate MD token type input}
\usage{
validate_token_type(token_type)
}
\arguments{
\item{token_type}{character vector either read_write or read_scaling}
}
\value{
vector
}
\description{
Validate MD token type input
}
