% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/minimumDistanceMethod.R
\name{xmm}
\alias{xmm}
\title{Get number order vector which is used in md}
\usage{
xmm(j)
}
\arguments{
\item{j}{integer}
}
\value{
vector related to number order
}
\description{
Get number order vector which is used in md
}

