% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/minimumDistanceMethod.R
\name{xym}
\alias{xym}
\title{Get number order matrix which is used in md}
\usage{
xym(i, length)
}
\arguments{
\item{i}{integer}

\item{length}{the number of grids in domain}
}
\value{
matrix related to number order
}
\description{
Get number order matrix which is used in md
}

