% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-md.R
\name{cd}
\alias{cd}
\title{Change Database}
\usage{
cd(.con, database_name, schema_name)
}
\arguments{
\item{.con}{connection}

\item{database_name}{database name}

\item{schema_name}{schema name}
}
\value{
message
}
\description{
Change Database
}
