% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{create_md_user}
\alias{create_md_user}
\title{Create a new MotherDuck user}
\usage{
create_md_user(user_name, motherduck_token = "MOTHERDUCK_TOKEN")
}
\arguments{
\item{user_name}{A character string specifying the MotherDuck user name whose
tokens should be listed.}

\item{motherduck_token}{Character. Either the name of an environment variable containing your
MotherDuck access token (default \code{"MOTHERDUCK_TOKEN"}) or the token itself.}
}
\value{
A tibble summarizing the API response, typically containing the newly created
username and associated metadata.
}
\description{
Sends a \code{POST} request to the MotherDuck REST API to create a new user
within your organization. This operation requires administrative privileges
and a valid access token.
}
\details{
This function calls the
\href{https://motherduck.com/docs/sql-reference/rest-api/users-create/}{MotherDuck Users API}
endpoint to create a new user under the authenticated account.
The provided token must belong to a user with permissions to manage
organization-level accounts.
}
\examples{
\dontrun{
# Create a new user in MotherDuck using an admin token stored in an environment variable
create_md_user("test_20250913", "MOTHERDUCK_TOKEN")
}

}
\seealso{
\code{\link[=delete_md_user]{delete_md_user()}} for deleting users, and
\code{\link[=list_md_user_tokens]{list_md_user_tokens()}} for listing tokens associated with a given user.
}
