% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-md.R
\name{create_or_replace_database}
\alias{create_or_replace_database}
\title{Overwrite or append tibble to motherduck database}
\usage{
create_or_replace_database(
  .data,
  .con,
  database_name,
  schema_name,
  table_name,
  write_type = "overwrite"
)
}
\arguments{
\item{.data}{tibble}

\item{.con}{duckdb connection}

\item{database_name}{name of database}

\item{schema_name}{name of schema}

\item{table_name}{name of table}

\item{write_type}{overwrite or append}
}
\value{
nothing
}
\description{
Overwrite or append tibble to motherduck database
}
