% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-db.R
\name{create_schema}
\alias{create_schema}
\title{Overwrite or append tibble to motherduck database}
\usage{
create_schema(.con, database_name, schema_name)

create_schema(.con, database_name, schema_name)
}
\arguments{
\item{.con}{duckdb connection}

\item{database_name}{name of database}

\item{schema_name}{name of schema}

\item{.data}{tibble}

\item{table_name}{name of table}

\item{write_type}{overwrite or append}
}
\value{
nothing

nothing
}
\description{
Overwrite or append tibble to motherduck database

Overwrite or append tibble to motherduck database
}
