% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{delete_md_user}
\alias{delete_md_user}
\title{Title}
\usage{
delete_md_user(user_name, motherduck_token)
}
\arguments{
\item{user_name}{motherduck user name}

\item{motherduck_token}{motherduck token or environment variable}
}
\description{
Title
}
