% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{describe_share}
\alias{describe_share}
\title{Describe share}
\usage{
describe_share(.con, share_name)
}
\arguments{
\item{.con}{A valid \code{DBI} connection (DuckDB / MotherDuck).}

\item{share_name}{shared path name}
}
\value{
tibble
}
\description{
Describe share
}
