% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{drop_share}
\alias{drop_share}
\title{Drop a MD share name}
\usage{
drop_share(.con, share_name)
}
\arguments{
\item{.con}{A valid \code{DBI} connection (DuckDB / MotherDuck).}

\item{share_name}{Share name}
}
\value{
message
}
\description{
Drop a MD share name
}
