% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-md.R
\name{list_extensions}
\alias{list_extensions}
\title{List motherduck extensions, their description and status}
\usage{
list_extensions(.con)
}
\arguments{
\item{.con}{DuckDB connection}
}
\value{
tibble
}
\description{
Lists available DuckDB extensions, their description, load / installed status and more
}
\examples{
\dontrun{
con <- DBI::dbConnect(duckdb::duckdb())
list_extensions(con)
}
}
