% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{list_shared_with_me_shares}
\alias{list_shared_with_me_shares}
\title{List all shares that are shared with you}
\usage{
list_shared_with_me_shares(.con)
}
\arguments{
\item{.con}{A valid \code{DBI} connection (DuckDB / MotherDuck).}
}
\value{
tibble
}
\description{
List all shares that are shared with you
}
