% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fns.R
\name{read_excel_duckdb}
\alias{read_excel_duckdb}
\title{Title}
\usage{
read_excel_duckdb(
  .con,
  to_database_name,
  to_schema_name,
  to_table_name,
  file_path,
  header,
  sheet,
  all_var_char,
  ignore_errors,
  range,
  stop_at_empty,
  empty_as_varchar
)
}
\arguments{
\item{empty_as_varchar}{}
}
\description{
Title
}
