% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{show_current_user}
\alias{show_current_user}
\title{Show current user}
\usage{
show_current_user(.con, motherduck_token = "MOTHERDUCK_TOKEN", return = "msg")
}
\arguments{
\item{.con}{connection}

\item{motherduck_token}{motherduck token}

\item{return}{msg or arg}
}
\value{
tibble or print statement
}
\description{
Show current user
}
