% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{show_current_user}
\alias{show_current_user}
\title{Show current user}
\usage{
show_current_user(.con, motherduck_token, return = "msg")
}
\arguments{
\item{.con}{A valid \code{DBI} connection (DuckDB / MotherDuck).}

\item{motherduck_token}{Character. Either the name of an environment variable containing your
MotherDuck access token (default \code{"MOTHERDUCK_TOKEN"}) or the token itself.}

\item{return}{msg or arg}
}
\value{
tibble or print statement
}
\description{
Show current user
}
