% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-md.R
\name{summary.tbl_lazy}
\alias{summary.tbl_lazy}
\title{Summarize for DBI objects}
\usage{
\method{summary}{tbl_lazy}(object, ...)
}
\arguments{
\item{object}{dbi object}

\item{...}{addtional argments, unused}
}
\value{
DBI object
}
\description{
Summarize for DBI objects
}
